requirements.txt
Flask==2.2.3
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.4.0
keras==2.12.0
tensorflow==2.12.0